refactor: moved dual agent to rai_bench
